Admin functionality added

+++ src/main/java/dao/impl/mysql/MySqlPaymentDao.java
@@ -110,7 +110,7 @@ public class MySqlPaymentDao implements PaymentDao {
     public Payment insert(Payment obj) {
         Objects.requireNonNull(obj);
 
-        int id = defaultDao.
+        int id = (int)defaultDao.
                 executeInsertWithGeneratedPrimaryKey(
                 INSERT,
                         obj.getAmount(),
